refactor(edge): isolate metrics endpoint service and state view
